refactor : Transtional and type matching
서비스단에서 CUD처리하는 곳을 transtional처리 해주었고 type 반환에 문제가 될 수 있는 부분을 처리해주었습니다.

+++ backend/src/main/java/com/ssafy/backend/api/controller/ApplicantStateController.java
@@ -104,7 +104,7 @@ public class ApplicantStateController {
 
             for (ApplicantState as: list
                  ) {
-                arrayList.add(ApplicantStateRes.builder().debate_topic(as.getDebateBoard().getDebateTopic()).current_applicant(applicantService.countCurrentApplicantByBoardNo(as.getDebateBoard().getBoardNo()))
+                arrayList.add(ApplicantStateRes.builder().debate_topic(as.getDebateBoard().getDebateTopic()).current_applicant((int) applicantService.countCurrentApplicantByBoardNo(as.getDebateBoard().getBoardNo()))
                         .max_applicant(as.getDebateBoard().getMaxApplicant()).accept(as.getAccept())
                         .board_finished(as.getDebateBoard().isBoardFinished()).build());
             }
@@ -137,7 +137,7 @@ public class ApplicantStateController {
 
             for (ApplicantState as: list
             ) {
-                arrayList.add(ApplicantStateRes.builder().debate_topic(as.getDebateBoard().getDebateTopic()).current_applicant(applicantService.countCurrentApplicantByBoardNo(as.getDebateBoard().getBoardNo()))
+                arrayList.add(ApplicantStateRes.builder().debate_topic(as.getDebateBoard().getDebateTopic()).current_applicant((int) applicantService.countCurrentApplicantByBoardNo(as.getDebateBoard().getBoardNo()))
                         .max_applicant(as.getDebateBoard().getMaxApplicant()).accept(as.getAccept())
                         .board_finished(as.getDebateBoard().isBoardFinished()).build());
             }
